Renamed gbmusicremote to gbmediaremote, fixed bug where any interaction with the button would cause a media play/pause toggle to be issued.
