+++ apps/gbmediaremote/ChangeLog
@@ -1,0 +1,1 @@
+0.01: Initial version
