+++ apps/gbmediaremote/icon.js
@@ -1,0 +1,1 @@
+require("heatshrink").decompress(atob("mEwwYxukmACOZZkwVIgmQCLxHopIRvdKg1PCMShkGoIABI54RgI/4AmyUBkkSoA1ECNYAPhMgEZ4RjAEZZ/YpwRwAGA="))
