--- apps/gbmusicremote/icon.js
@@ -1,1 +1,0 @@
-require("heatshrink").decompress(atob("kEgoX/AAQOHgoHOC5EUqAHMC5NRoAHLI5IXGA4wXJgIXGA4oXdKgNFA5h0JB4gHHC74/CKgqkOABwA=="))
